+++ clientele/graphql/__init__.py
@@ -1,0 +1,3 @@
+from clientele.graphql.client import GraphQLClient
+
+__all__ = ["GraphQLClient"]